+++ packages/effect-app/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @effect-app/prelude
 
+## 4.0.0-beta.182
+
 ## 4.0.0-beta.181
 
 ### Patch Changes